requirements.txt
attrs==23.2.0
certifi==2023.11.17
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
PySocks==1.7.1
python-dateutil==2.8.2
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0

allure-pytest==         2.13.2
allure-python-commons ==2.13.2
astroid              == 3.0.1
asttokens          ==   2.4.1
attrs               ==  23.2.0
certifi             ==  2023.11.17
cffi                ==  1.16.0
charset-normalizer  ==  3.3.2
colorama            ==  0.4.6
cowsay              ==  6.1
ddt                 ==  1.7.1
decorator           ==  5.1.1
dill                ==  0.3.7
et-xmlfile          ==  1.1.0
exceptiongroup      ==  1.2.0
executing           ==  2.0.1
h11                 ==  0.14.0
idna                ==  3.6
iniconfig           ==  2.0.0
ipython             ==  8.17.2
isort               ==  5.12.0
jedi                ==  0.19.1
matplotlib-inline   ==  0.1.6
mccabe              ==  0.7.0
mypy                ==  1.7.1
mypy-extensions     ==  1.0.0
openpyxl            ==  3.1.2
outcome             ==  1.3.0.post0
packaging           ==  23.2
parso                == 0.8.3
pip                  == 24.0
platformdirs         == 4.0.0
pluggy               == 1.3.0
prompt-toolkit       == 3.0.40
pure-eval             ==0.2.2
pycparser           ==  2.21
Pygments             == 2.16.1
pylint               == 3.0.2
PySocks              == 1.7.1
pytest              ==  8.0.1
python-dateutil     ==  2.8.2
requests            ==  2.31.0
selenium            ==  4.16.0
setuptools          ==  68.2.2
six                 ==  1.16.0
sniffio             ==  1.3.0
sortedcontainers    ==  2.4.0
stack-data          ==  0.6.3
syspy                == 0.2
tomli               ==  2.0.1
tomlkit             ==  0.12.3
traitlets            == 5.13.0
trio                 == 0.24.0
trio-websocket       == 0.11.1
typing_extensions    == 4.8.0
urllib3              == 2.1.0
wcwidth              == 0.2.9
wheel               ==  0.41.2
wsproto             ==  1.2.0


